feat(secrets): use os-backed root custody
