+++ backend/crates/mcpmate-secrets/Cargo.toml
@@ -7,4 +7,29 @@ description = "Runtime secret references and provider boundaries for MCPMate"
 authors = ["Loocor <loocor@gmail.com>"]
 
 [dependencies]
+anyhow = "1.0"
+base64 = "0.22.1"
+ring = "0.17.14"
+serde = { version = "1.0", features = ["derive"] }
+sha2 = "0.10"
+sqlx = { version = "0.8.1", features = [
+  "runtime-tokio",
+  "sqlite",
+  "chrono",
+  "runtime-tokio-rustls",
+] }
 thiserror = "1.0"
+
+[dev-dependencies]
+serial_test = "2.0"
+tempfile = "3.2"
+tokio = { version = "1.44", features = ["full"] }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+keyring = { version = "3.6.3", default-features = false, features = ["apple-native"] }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+keyring = { version = "3.6.3", default-features = false, features = ["windows-native"] }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+keyring = { version = "3.6.3", default-features = false, features = ["sync-secret-service", "crypto-rust"] }
