+++ backend/src/api/handlers/secrets.rs
@@ -110,10 +110,12 @@ pub async fn list_secret_usages(
 }
 
 fn get_secret_store(state: &Arc<AppState>) -> Result<Arc<crate::core::secrets::store::LocalSecretStore>, ApiError> {
-    state
-        .secret_store
-        .clone()
-        .ok_or_else(|| ApiError::InternalError("Secret store is unavailable".to_string()))
+    state.secret_store.clone().ok_or_else(|| {
+        ApiError::InternalError(
+            "Secret store is unavailable. Unlock or configure the operating-system secure storage provider."
+                .to_string(),
+        )
+    })
 }
 
 fn map_secret_store_error(error: anyhow::Error) -> ApiError {
